Solve Flipped Place Puzzle
My teacher gave us this puzzle about two weeks before school ended. The string at the beginning is the puzzle we were supposed to solve. The answer to each letter is the place of the letter in the alphabet but flipped. For example, if you see a 'z', it is actually an 'a'. If you see a 'c', it is an 'x', and so on.